+++ src/main/java/lotto/view/InputView.java
@@ -12,15 +12,15 @@ public class InputView {
     private InputView() {
     }
 
-    public static int inputMoney() {
+    public static long inputMoney() {
         System.out.println("구입금액을 입력해 주세요.");
-        return Integer.parseInt(scanner.nextLine());
+        return Long.parseLong(scanner.nextLine());
     }
 
     public static List<Integer> inputWinningNumbers() {
         System.out.println("지난 주 당첨 번호를 입력해 주세요.");
         String[] values = scanner.nextLine().split(",");
-        List<Integer> winningNumbers =  Arrays.stream(values)
+        List<Integer> winningNumbers = Arrays.stream(values)
             .map(value -> Integer.parseInt(value.trim()))
             .collect(Collectors.toList());
         System.out.println();
